Factor function service registrations
