+++ Predictorator.Tests/Predictorator.Tests.csproj
@@ -23,6 +23,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Predictorator\Predictorator.csproj" />
+    <ProjectReference Include="..\Predictorator.Functions\Predictorator.Functions.csproj" />
   </ItemGroup>
 
   <ItemGroup>
